Move bungee classes to separate module
