+++ triton-bungeecord/src/main/java/com/rexcantor64/triton/bungeecord/commands/handler/BungeeSender.java
@@ -1,6 +1,8 @@
-package com.rexcantor64.triton.commands.handler;
+package com.rexcantor64.triton.bungeecord.commands.handler;
 
 import com.rexcantor64.triton.Triton;
+import com.rexcantor64.triton.commands.handler.NoPermissionException;
+import com.rexcantor64.triton.commands.handler.Sender;
 import lombok.AllArgsConstructor;
 import lombok.val;
 import net.md_5.bungee.api.ChatColor;
