+++ triton-bungeecord/src/main/java/com/rexcantor64/triton/bungeecord/packetinterceptor/BungeeDecoder.java
@@ -1,7 +1,7 @@
-package com.rexcantor64.triton.packetinterceptor;
+package com.rexcantor64.triton.bungeecord.packetinterceptor;
 
 import com.rexcantor64.triton.Triton;
-import com.rexcantor64.triton.player.BungeeLanguagePlayer;
+import com.rexcantor64.triton.bungeecord.player.BungeeLanguagePlayer;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.handler.codec.MessageToMessageDecoder;
 import net.md_5.bungee.protocol.PacketWrapper;
