+++ triton-bungeecord/src/main/java/com/rexcantor64/triton/bungeecord/plugin/BungeePlugin.java
@@ -1,13 +1,14 @@
-package com.rexcantor64.triton.plugin;
+package com.rexcantor64.triton.bungeecord.plugin;
 
 import com.rexcantor64.triton.BungeeMLP;
 import com.rexcantor64.triton.Triton;
 import com.rexcantor64.triton.logger.JavaLogger;
 import com.rexcantor64.triton.logger.TritonLogger;
-import com.rexcantor64.triton.terminal.BungeeTerminalManager;
+import com.rexcantor64.triton.plugin.Platform;
+import com.rexcantor64.triton.plugin.PluginLoader;
+import com.rexcantor64.triton.bungeecord.terminal.BungeeTerminalManager;
 import com.rexcantor64.triton.terminal.Log4jInjector;
 import net.md_5.bungee.api.plugin.Plugin;
-import org.apache.logging.log4j.core.config.plugins.util.PluginType;
 
 import java.util.logging.Level;
