+++ triton-bungeecord/src/main/java/com/rexcantor64/triton/bungeecord/terminal/BungeeTerminalManager.java
@@ -1,4 +1,4 @@
-package com.rexcantor64.triton.terminal;
+package com.rexcantor64.triton.bungeecord.terminal;
 
 import net.md_5.bungee.BungeeCord;
 import net.md_5.bungee.log.ColouredWriter;
