+++ triton-bungeecord/src/main/resources/config_bungeecord.yml
